fixed komplettering to lab2

+++ src/standalone/lab01/about-main.c
@@ -7,8 +7,8 @@ int main( int argc, char *argv[] )
   int maxwidth = 21;
   for (int i = 0 ; i < argc ; ++i)
   {
-    int width = maxwidth - (int)strlen(argv[i]);
-    printf("%s %*i \n", argv[i],width, (int)strlen(argv[i]) );
+    int width = maxwidth - strlen(argv[i]);
+    printf("%s %*i \n", argv[i],width, (int)strlen(argv[i]));
     summation = (int)strlen(argv[i]) + summation;
   }
   int width = maxwidth - (int)strlen("Total length");
